requirements.txt
tensorflow==2.13.0
streamlit==1.24.1
protobuf==3.20.3
pandas
pillow
numpy